Add ActionUIPropertyValidation protocol for non-View elements
Added ActionUIRegistry.validationRegistrations and registerAdditionalElementValidations() to register new elements: WindowGroup, CommandGroup and CommandMenu which are not views and don't conform to ActionUIViewConstruction protocol.
That way the property validation happens early, the same way as for views, not during element construction.
registry.validateProperties(forElementType:, properties:) now calls View.validateProperties() internally for views and skips it for non-views