Initialize ViewModel states before construction
Some view maintain "states" storage, which is a published property of ViewModel. Because it is published it cannot be mutated during view construction in 'buildView:' or Combine will enter infinite recursion.
The design is to implement 'initialStates:' closure similar to 'initialValue:' for specific views which need it. WindowModel.populateViewModels() will call these functions for each element