Change picker to use tags to identify items
Previously we used the text of menu/chooser element item to be an identifying tag but this is not great, especially if the string can be localizable. A tag is effectively a unique identifier chosen by developer. If the UI designer does not provide explicit item tags, we use 1-based index (as string) for tag, detaching it from the actual name (client code must keep track of index and the item it represents)