Binding.set fix
User initiated value changes in controls trigger Binding.set callback. But what seems to be a terrible design decision on Apple side, that may happen during SsiftUI.View rendering. The result is that mutating `model.value` triggers a runtime SwiftUI warning:
"Publishing changes from within view updates is not allowed, this will cause undefined behavior."

We cannot wrap the model updates in a `Task { @MainActor in }` because it does not dispatch when called on the main thread. The only reliable method is to use `DispatchQueue.main.async { }`
Another issue is that Binding.set may be called even if the value is not changing. In order to avoid spurious model updates and triggering valueChangeActionID action handlers we guard against same values first.